Building rational transfer functions

RationalTransferFunction Construction of rational transfer function object.

rtf Building rational transfer functions.

DescriptorSystems.RationalTransferFunction — Type

 RationalTransferFunction{T,X}(num::Polynomial{T,X}, den::Polynomial{T,X}, Ts::Real) where T <: Number

Construct a rational transfer function model from its numerator and denominator polynomials num and den, respectively, and a sampling time Ts.

If r::RationalTransferFunction{T,X} is a rational transfer function system model object defined as r(λ) = num(λ)/den(λ), where num(λ) and den(λ) are polynomials in the indeterminate λ, then:

r.num is the numerator polynomial num(λ) with coefficients of type T and indeterminate λ;

r.den is the denominator polynomial den(λ) with coefficients of type T and indeterminate λ;

r.Ts is the sampling time Ts, where Ts can have the following values:

Ts = 0 for a continuous-time system and λ = s is the complex variable in the Laplace transform;

Ts > 0 or Ts = -1 for a discrete-time system and λ = z is the complex variable in the Z-transform; Ts = -1 indicates a discrete-time system with an unspecified sampling time;

The symbol (or variable) used for the indeterminate λ is the common symbol used for the indeterminates of the polynomials num(λ) and den(λ) and can be obtained as r.var. The roots of the numerator polynomial num(λ) (also called zeros of r(λ)) can be obtained as r.zeros, while the roots of the denominator polynomial den(λ) (also called poles of r(λ)) can be obtained as r.poles. The ratio of the leading polynomial coefficients of num(λ) and den(λ) (also called gain of r(λ)) can be obtained as r.gain.

DescriptorSystems.rtf — Function

r = rtf(num, den; Ts = 0, var = Polynomials.indeterminate(num))

Create for the polynomials num(λ) and den(λ), sampling time Ts and variable name var the rational transfer function r(λ) = num(λ)/den(λ) of a single-input single-output system of the form

Y(λ) = r(λ) U(λ),

where U(λ) and Y(λ) are the Laplace transformed, if Ts = 0, or Z-transformed, if Ts ≠ 0, system input u(t) and system output y(t), respectively, and λ = s, if Ts = 0, or λ = z, if Ts ≠ 0. The resulting r is such that r.Ts = Ts and r.var = var.

Both num(λ) and den(λ) can be real or complex numbers as well.

r = rtf(f; Ts = f.Ts, var = f.var)

Set for the rational transfer function r(λ) = f(λ) the sampling time to Ts and variable name to var such that r(λ) = f(λ), r.Ts = Ts (default r.Ts = f.Ts) and r.var = var (default r.var = f.var).

r = rtf(p; Ts = 0, var = Polynomials.indeterminate(p))

Create for the polynomial p(λ), sampling time Ts and variable var the rational transfer function r(λ) = p(λ) such that r.Ts = Ts (default r.Ts = 0) and r.var = var (default (default r.var = Polynomials.indeterminate(p)). p(λ) can be a real or complex number as well.

r = rtf(var; Ts = 0.)
r = rtf('s'; Ts = 0.) or r = rtf('z'; Ts = -1.) 
r = rtf("s"; Ts = 0.) or r = rtf("z"; Ts = -1.) 
r = rtf(:s; Ts = 0.) or r = rtf(:z; Ts = -1.)

Create the rational transfer function r(λ) = λ, such that r.var and r.Ts are set as follows: (1) r.var = :s and r.Ts = 0 if var = 's', or var = "s" or var = :s ; (2) r.var = :z and r.Ts = Ts if var = 'z', or var = "z" or var = :z; (3) r.var = var and r.Ts = Ts (default Ts = 0.) otherwise.

r = rtf(z, p, k; Ts = 0, var = :s)

Create from the roots (zeros) z, poles p, gain k, sampling time Ts and variable name var the rational transfer function r(λ) = k*num(λ)/den(λ), where num(λ) and den(λ) are monic polynomials with roots equal z and p, respectively, and such that r.Ts = Ts (default r.Ts = 0) and r.var = var (default r.var = :s if Ts = 0 or r.var = :z if Ts ≠ 0).